updated strategy part 2

+++ sw/airborne/modules/computer_vision/corner_detection.h
@@ -32,7 +32,7 @@
 
 
 struct flow {
-    int counter;
+    float counter;
     float total;
     float average;
 };
